pyproject.toml
[project]
name = "team-6-recommender-system"
version = "0.1.0"
description = "Default template for PDM package"
authors = [
    {name = "Team 6", email = "team6@gmail.com"},
]
dependencies = [
    "pandas>=2.0.0",
    "numpy==1.24.0",
    "scikit-learn==1.3.0",
    "xgboost>=2.0.0",
    "joblib>=1.3.0",
    "requests>=2.31.0",
    "confluent-kafka>=2.2.0",
    "huggingface-hub",
    "tqdm>=4.66.0",
    "flask>=3.0.0",
    "kafka-python>=2.0.0",
    "tensorflow-cpu>=2.17.0",
    "psutil>=5.9.0",
    "scipy>=1.11.0",
    "scikit-surprise>=1.1.4",
    "pytest>=7.4.0",
    "pytest-cov>=4.1.0",
    "implicit==0.7.2",
]
requires-python = "==3.10.*"
readme = "README.md"
license = {text = "MIT"}


[tool.pdm]
source-includes = ["src"]

[tool.pdm.build]
package-dir = "src"

[dependency-groups]
dev = [
    "pytest>=8.4.2",
    "pytest-cov>=7.0.0",
]
